move all package objects into conventional locations
By usual convention, scala package objects reside in a directory named
after the package, in a file called package.scala.  IntelliJ's scala
plugin gets very confused when this convention is not followed and fails
to find the classes from the package when used elsewhere (e.g. in tests)

Moving the files to match convention appears to fix the IntelliJ issues.